Rename utils.py to be compatible with ComfyUI in embeddable python env
