+++ service/aipg_utils.py
